Rename : ErrorMessage 파일 위치 domain 밖으로 이동
